# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: Mishwar - Complete ride hailing and group trips platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /rides:
    get:
      operationId: listRides
      tags: [rides]
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, accepted, in_progress, completed, cancelled]
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RideList"
    post:
      operationId: createRide
      tags: [rides]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRideBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ride"
  /rides/{rideId}:
    get:
      operationId: getRide
      tags: [rides]
      parameters:
        - name: rideId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ride"
    patch:
      operationId: updateRide
      tags: [rides]
      parameters:
        - name: rideId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRideBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ride"

  # Group Trips
...
  /invite-trips:
    get:
      operationId: listInviteTrips
      tags: [inviteTrips]
      summary: List my invite trips
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InviteTripList"
    post:
      operationId: createInviteTrip
      tags: [inviteTrips]
      summary: Create an invite trip
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInviteTripBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InviteTrip"
  /invite-trips/{inviteTripId}:
    get:
      operationId: getInviteTrip
      tags: [inviteTrips]
      parameters:
        - name: inviteTripId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InviteTrip"
    patch:
      operationId: updateInviteTrip
      tags: [inviteTrips]
      parameters:
        - name: inviteTripId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateInviteTripBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InviteTrip"
  /invite/{inviteCode}:
    get:
      operationId: getInviteTripByCode
      tags: [inviteTrips]
      summary: Get trip by invite code (public)
      parameters:
        - name: inviteCode
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InviteTrip"
  /invite/{inviteCode}/respond:
    post:
      operationId: respondToInvite
      tags: [inviteTrips]
      summary: Accept or decline an invite
      parameters:
        - name: inviteCode
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RespondToInviteBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InviteGuest"
  /invite-trips/{inviteTripId}/confirm:
    post:
      operationId: confirmInviteTrip
      tags: [inviteTrips]
      summary: Confirm and book the invite trip
      parameters:
        - name: inviteTripId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InviteTrip"

  # Rentals
  /rentals:
    get:
      operationId: listRentals
      tags: [rentals]
      summary: List available rental cars
      parameters:
        - name: vehicleType
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RentalCarList"
    post:
      operationId: createRentalBooking
      tags: [rentals]
      summary: Book a rental car
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRentalBookingBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RentalBooking"
...
  /rentals/{rentalId}:
    get:
      operationId: getRentalBooking
      tags: [rentals]
      parameters:
        - name: rentalId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RentalBooking"
    delete:
      operationId: cancelRentalBooking
      tags: [rentals]
      parameters:
        - name: rentalId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RentalBooking"

  # Driver Bookings
...
components:
  schemas:
...
    Driver:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        phone:
          type: string
        vehicleMake:
          type: string
        vehicleModel:
          type: string
        vehicleYear:
          type: integer
        vehiclePlate:
          type: string
        vehicleType:
          type: string
          enum: [economy, comfort, xl, bus, van, minibus]
        rating:
          type: number
          nullable: true
        totalTrips:
          type: integer
        isAvailable:
          type: boolean
        driverStatus:
          type: string
          enum: [pending, approved, rejected, suspended]
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, userId, name, phone, vehicleMake, vehicleModel, vehicleYear, vehiclePlate, vehicleType, totalTrips, isAvailable, createdAt]
...
    Ride:
      type: object
      properties:
        id:
          type: integer
        passengerId:
          type: integer
        driverId:
          type: integer
          nullable: true
        fromAddress:
          type: string
        toAddress:
          type: string
        fromLat:
          type: number
          nullable: true
        fromLng:
          type: number
          nullable: true
        toLat:
          type: number
          nullable: true
        toLng:
          type: number
          nullable: true
        status:
          type: string
          enum: [pending, accepted, in_progress, completed, cancelled]
        estimatedPrice:
          type: number
          nullable: true
        finalPrice:
          type: number
          nullable: true
        vehicleType:
          type: string
          enum: [economy, comfort, xl]
        scheduledAt:
          type: string
          format: date-time
          nullable: true
        completedAt:
          type: string
          format: date-time
          nullable: true
        notes:
          type: string
          nullable: true
        driver:
          $ref: "#/components/schemas/Driver"
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, passengerId, fromAddress, toAddress, status, vehicleType, createdAt]
    RideList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Ride"
        total:
          type: integer
      required: [items, total]
    CreateRideBody:
      type: object
      properties:
        fromAddress:
          type: string
        toAddress:
          type: string
        fromLat:
          type: number
        fromLng:
          type: number
        toLat:
          type: number
        toLng:
          type: number
        vehicleType:
          type: string
          enum: [economy, comfort, xl]
          default: economy
        scheduledAt:
          type: string
          format: date-time
        notes:
          type: string
      required: [fromAddress, toAddress]
    UpdateRideBody:
      type: object
      properties:
        status:
          type: string
          enum: [accepted, in_progress, completed, cancelled]
        driverId:
          type: integer
        finalPrice:
          type: number
...
    InviteTrip:
      type: object
      properties:
        id:
          type: integer
        creatorId:
          type: integer
        title:
          type: string
        fromAddress:
          type: string
        toAddress:
          type: string
        vehicleType:
          type: string
          enum: [economy, comfort, xl, bus, van, minibus]
        scheduledAt:
          type: string
          format: date-time
        inviteCode:
          type: string
        status:
          type: string
          enum: [open, confirmed, completed, cancelled]
        maxGuests:
          type: integer
        notes:
          type: string
          nullable: true
        estimatedPrice:
          type: number
          nullable: true
        guests:
          type: array
          items:
            $ref: "#/components/schemas/InviteGuest"
        createdAt:
          type: string
          format: date-time
      required: [id, creatorId, title, fromAddress, toAddress, vehicleType, scheduledAt, inviteCode, status, maxGuests, createdAt]
    InviteTripList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/InviteTrip"
        total:
          type: integer
      required: [items, total]
    InviteGuest:
      type: object
      properties:
        id:
          type: integer
        inviteTripId:
          type: integer
        name:
          type: string
        phone:
          type: string
        pickupAddress:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, accepted, declined]
        createdAt:
          type: string
          format: date-time
      required: [id, inviteTripId, name, phone, status, createdAt]
    CreateInviteTripBody:
      type: object
      properties:
        title:
          type: string
        fromAddress:
          type: string
        toAddress:
          type: string
        vehicleType:
          type: string
          enum: [economy, comfort, xl, bus, van, minibus]
        scheduledAt:
          type: string
          format: date-time
        maxGuests:
          type: integer
          default: 4
        notes:
          type: string
      required: [title, fromAddress, toAddress, vehicleType, scheduledAt]
    UpdateInviteTripBody:
      type: object
      properties:
        title:
          type: string
        scheduledAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [confirmed, cancelled]
        notes:
          type: string
    RespondToInviteBody:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        pickupAddress:
          type: string
        response:
          type: string
          enum: [accepted, declined]
      required: [name, phone, response]

    # Rentals
    RentalCar:
      type: object
      properties:
        id:
          type: integer
        make:
          type: string
        model:
          type: string
        year:
          type: integer
        plate:
          type: string
        vehicleType:
          type: string
          enum: [economy, comfort, xl, suv, luxury]
        dailyRate:
          type: number
        isAvailable:
          type: boolean
        imageUrl:
          type: string
          nullable: true
        features:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, make, model, year, plate, vehicleType, dailyRate, isAvailable, createdAt]
    RentalCarList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/RentalCar"
        total:
          type: integer
      required: [items, total]
    RentalBooking:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        carId:
          type: integer
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        totalDays:
          type: integer
        totalPrice:
          type: number
        status:
          type: string
          enum: [pending, confirmed, active, returned, cancelled]
        car:
          $ref: "#/components/schemas/RentalCar"
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, userId, carId, startDate, endDate, totalDays, totalPrice, status, createdAt]
...
    CreateRentalBookingBody:
      type: object
      properties:
        carId:
          type: integer
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        notes:
          type: string
      required: [carId, startDate, endDate]

    # Driver Bookings
...